python/requirements.txt
pandas
matplotlib
scikit-learn
yfinance
python-telegram-bot
seaborn
requests
json
